+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:color_tile/screen/playing_page.dart';
 import 'package:color_tile/screen/result_page.dart';
 import 'package:flutter/material.dart';
-import 'package:color_tile/screen/homepage.dart';
+import 'package:color_tile/screen/title_page.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:go_router/go_router.dart';
